+++ src/main/java/com/bootcamp/pragma/stockmicroservice/domain/spi/IBrandPersistencePort.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface IBrandPersistencePort {
     void saveMarca(Brand brand);
     List<Brand> getAllMarca(int page, String order);
-    Brand getMarcaById(int id);
+    Brand getMarcaById(Long id);
 }
